serde: bench serialization of coads
